pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.66.3",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "pypdf2>=3.0.1",
    "pytz>=2025.1",
    "reportlab>=4.3.1",
]
